+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "A child's play".
 - Tests for "Sum of spiral's diagonals".
 - Tests for "Brackets, extreme edition".
+- Tests for "The electrician apprentice".
 
 ## [1.6.0] - 2022-03-11
 ### Added